Add fee rebates to reward to incentivize trading
- Track prev_fees for fee delta calculation
- Reward = realized_delta + unrealized_delta + fee_delta
- With maker_fee=-0.25bps, each trade earns rebate
- Encourages market participation without reckless spreads

+++ litepool/rltrader/position.h
@@ -28,7 +28,7 @@ namespace RLTrader {
         double averagePrice = 0.0;
         double netAmount = 0.0;
         double totalFee = 0.0;
-        int numOfTrades = 0;
+        long numOfTrades = 0;
         double initialBalance = 0.0;
         double balance = 0.0;
         TradeInfo trade_info;
